requirements.txt
asgiref==3.6.0
Django==4.1.4
django-auth==0.1.9
django-crispy-forms==1.14.0
django-haystack==3.2.1
django-localflavor==3.1
djangorestframework==3.14.0
Pillow==9.3.0
python-decouple==3.6
python-stdnum==1.18
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
Whoosh==2.7.4
